Initial Amiga support, refactoring

+++ showet.py
@@ -1,8 +1,9 @@
 #!/usr/bin/env python3
 import urllib.request, json
-import os.path
-from os import listdir
+import os
 import sys
+from platformwindows import PlatformWindows
+from platformamiga import PlatformAmiga
 
 # Note: This is still early proof of concept, but seems to work.
 
@@ -23,7 +24,6 @@ prod_download_url = None
 prod_download_filename = None
 prod_json = None
 prod_json_filename = datadir + "/pouet.json"
-wineprefix = os.getcwd() + '/wineprefix'
 
 print("Prod URL: " + prod_url)
 
@@ -41,21 +41,50 @@ else:
         f.write(prod_json)
         f.close()
 
+print(prod_json)
+
 data = json.loads(prod_json)
+
+platforms = data['prod']['platforms'].values()
+
+# Just pick one:
+prod_platform = None
+for p in platforms:
+    prod_platform = p['slug']
+
+print("\tName: " + data['prod']['name'])
+try:
+    print("\tBy: " + data['prod']['groups'][0]['name'])
+except IndexError:
+    pass
+print("\tType: " + data['prod']['type'])
+print("\tReleased: " + data['prod']['releaseDate'])
+print("\tPlatform: " + prod_platform)
+
+platorm_runners = [PlatformAmiga(), PlatformWindows()]
+runner = None
+
+for r in platorm_runners:
+    if prod_platform in r.supported_platforms():
+        runner = r
+
+if not runner:
+    print("Platform " + prod_platform + " not supported (yet!).")
+    exit(-1)
 
 # Get necessary fields from the data
 
 prod_download_url = data['prod']['download']
 prod_download_url = prod_download_url.replace("https://files.scene.org/view", "https://files.scene.org/get")
 prod_download_filename = prod_download_url.split('/')
-prod_download_filename = prod_download_filename[len(prod_download_filename)-1]
+prod_download_filename = datadir + "/" +prod_download_filename[len(prod_download_filename)-1]
 
 # Download the prod if needed
 if not prod_download_filename:
     print("Error: couldn't get filename")
     exit(-1)
 
-if os.path.exists(prod_download_filename):
+if os.path.exists(datadir + "/_FILES_DOWNLOADED"):
     print("File already downloaded")
 else:
     print("Downloading prod file from " + prod_download_url + "...")
@@ -68,46 +97,26 @@ else:
     print("Downloaded ", prod_download_filename)
 
 
-# Decompress the file if needed
-if prod_download_filename.endswith(".zip"):
-    print("Unzipping", prod_download_filename)
-    ret = os.system("unzip -u -d " + datadir + " " + prod_download_filename)
-    if ret != 0:
-        print("Unzipping file failed!")
+    # Decompress the file if needed
+    if prod_download_filename.endswith(".zip"):
+        print("Unzipping", prod_download_filename)
+        ret = os.system("unzip -u -d " + datadir + " " + prod_download_filename)
+        if ret == 0:
+            # Delete the original file
+            os.remove(prod_download_filename)
+        else:
+            print("Unzipping file failed!")
 
-else:
-    print("TODO: Currently only zipped files supported..")
-    exit(-1)
+    if prod_download_filename.endswith(".lha"):
+        print("Extracting lha ", prod_download_filename)
+        ret = os.system("lha xw=" + datadir + " " + prod_download_filename)
+        if ret == 0:
+            # Delete the original file
+            os.remove(prod_download_filename)
+        else:
+            print("Unzipping file failed!")
 
-# Delete the original file
-os.remove(prod_download_filename)
+    open(datadir + "/_FILES_DOWNLOADED", 'a').close()
 
-# Guess which file to execute
-files = [f for f in listdir(datadir) if os.path.isfile(os.path.join(datadir, f))]
-
-exefiles = [f for f in files if (f.lower().endswith(".exe") or f.lower().endswith(".EXE")) ]
-
-if len(exefiles) == 0:
-    print("Didn't find any exe files.")
-    exit(-1)
-
-exefile = exefiles[0]
-
-print("Guessed executable file: " + exefile)
-
-exepath = datadir + "/" + exefile
-
-# Setup wine if needed
-
-os.putenv("WINEPREFIX", wineprefix)
-
-if not os.path.exists(wineprefix):
-    os.makedirs(wineprefix)
-    print("Creating wine prefix: " + str(wineprefix))
-    os.system('WINEARCH="win32" winecfg')
-
-# Run the demo
-
-print("Running " + exepath + "...")
-os.chdir(datadir)
-os.system('wine ' + exefile)
+runner.setup(datadir, prod_platform)
+runner.run()
